+++ java/org/ohdsi/sql/SqlTranslate.java
@@ -35,7 +35,7 @@ import java.util.regex.Pattern;
 
 public class SqlTranslate {
 	public static int SESSION_ID_LENGTH = 8;
-	public static int MAX_ORACLE_TABLE_NAME_LENGTH = 30;
+	public static int MAX_TABLE_NAME_LENGTH = 63; // PostreSQL default limit
 	private static Map<String, List<String[]>> targetToReplacementPatterns = null;
 	private static ReentrantLock lock = new ReentrantLock();
 	private static Random random = new Random();
@@ -604,13 +604,13 @@ public class SqlTranslate {
 		Matcher matcher = pattern.matcher(sql);
 		Set<String> longTempNames = new HashSet<String>();
 		while (matcher.find())
-			if (matcher.group().length() > MAX_ORACLE_TABLE_NAME_LENGTH - SESSION_ID_LENGTH - 1)
+			if (matcher.group().length() > MAX_TABLE_NAME_LENGTH - SESSION_ID_LENGTH - 1)
 				longTempNames.add(matcher.group());
 
 		for (String longName : longTempNames)
 			warnings.add("Temp table name '" + longName + "' is too long. Temp table names should be shorter than "
-					+ (MAX_ORACLE_TABLE_NAME_LENGTH - SESSION_ID_LENGTH)
-					+ " characters to prevent Oracle from crashing.");
+					+ (MAX_TABLE_NAME_LENGTH - SESSION_ID_LENGTH)
+					+ " characters to prevent some DMBSs from throwing an error.");
 
 		// normal table names:
 		pattern = Pattern.compile("(create|drop|truncate)\\s+table +[0-9a-zA-Z_]+");
@@ -618,12 +618,12 @@ public class SqlTranslate {
 		Set<String> longNames = new HashSet<String>();
 		while (matcher.find()) {
 			String name = sql.substring(matcher.start() + matcher.group().lastIndexOf(" "), matcher.end());
-			if (name.length() > MAX_ORACLE_TABLE_NAME_LENGTH && !longTempNames.contains("#" + name))
+			if (name.length() > MAX_TABLE_NAME_LENGTH && !longTempNames.contains("#" + name))
 				longNames.add(name);
 		}
 		for (String longName : longNames)
 			warnings.add("Table name '" + longName + "' is too long. Table names should be shorter than "
-					+ MAX_ORACLE_TABLE_NAME_LENGTH + " characters to prevent Oracle from crashing.");
+					+ MAX_TABLE_NAME_LENGTH + " characters to prevent some DMBSs from throwing an error.");
 
 		return warnings.toArray(new String[warnings.size()]);
 	}